requirements.txt
torch>=1.9.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.1
gradio>=3.0.0
Pillow>=8.0.0
numpy>=1.19.0